updated bundlestitchingservice logic, pulled in changes onto new branch
