+++ src/main/java/uk/gov/hmcts/reform/em/orchestrator/service/caseupdater/CcdBundleStitchingService.java
@@ -1,13 +1,9 @@
 package uk.gov.hmcts.reform.em.orchestrator.service.caseupdater;
 
-import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ArrayNode;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validator;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import uk.gov.hmcts.reform.em.orchestrator.service.ccdcallbackhandler.CcdCallbackDto;
@@ -19,59 +15,29 @@ import uk.gov.hmcts.reform.em.orchestrator.stitching.StitchingServiceException;
 import uk.gov.hmcts.reform.em.orchestrator.stitching.dto.CdamDto;
 import uk.gov.hmcts.reform.em.orchestrator.util.StringUtilities;
 
-import java.io.IOException;
-import java.util.List;
-import java.util.Optional;
 import java.util.Set;
-import java.util.Spliterator;
-import java.util.Spliterators;
-import java.util.stream.StreamSupport;
-
-import static pl.touk.throwing.ThrowingFunction.unchecked;
 
 @Service
 @Transactional
-public class CcdBundleStitchingService implements CcdCaseUpdater {
+public class CcdBundleStitchingService extends UpdateCase {
 
-    private final Logger logger = LoggerFactory.getLogger(CcdBundleStitchingService.class);
-
-    private final ObjectMapper objectMapper;
-    private final JavaType type;
     private final Validator validator;
     private final StitchingService stitchingService;
 
     public CcdBundleStitchingService(ObjectMapper objectMapper,
                                      StitchingService stitchingService,
                                      Validator validator) {
-        this.objectMapper = objectMapper;
+        super(objectMapper);
         this.stitchingService = stitchingService;
-        type = objectMapper.getTypeFactory().constructParametricType(CcdValue.class, CcdBundleDTO.class);
         this.validator = validator;
     }
 
-    @SuppressWarnings("java:S2139")
     @Override
     public JsonNode updateCase(CcdCallbackDto ccdCallbackDto) {
-        Optional<ArrayNode> maybeBundles = ccdCallbackDto.findCaseProperty(ArrayNode.class);
-
-        if (maybeBundles.isPresent()) {
-            List<JsonNode> newBundles = StreamSupport
-                    .stream(Spliterators.spliteratorUnknownSize(
-                            maybeBundles.get().iterator(), Spliterator.ORDERED), false)
-                    .parallel()
-                    .map(unchecked(this::bundleJsonToBundleValue))
-                    .map(bundle -> bundle.getValue().getEligibleForStitchingAsBoolean()
-                            ? this.stitchBundle(bundle, ccdCallbackDto) : bundle)
-                    .map(bundleDto -> objectMapper.convertValue(bundleDto, JsonNode.class))
-                    .toList();
-
-            maybeBundles.get().removeAll();
-            maybeBundles.get().addAll(CcdCaseUpdater.reorderBundles(newBundles, objectMapper, type));
-        }
-        return ccdCallbackDto.getCaseData();
+        return super.updateCase(ccdCallbackDto);
     }
 
-    private CcdValue<CcdBundleDTO> stitchBundle(CcdValue<CcdBundleDTO> bundle, CcdCallbackDto ccdCallbackDto) {
+    protected CcdValue<CcdBundleDTO> stitchBundle(CcdValue<CcdBundleDTO> bundle, CcdCallbackDto ccdCallbackDto) {
         bundle.getValue().setCoverpageTemplateData(ccdCallbackDto.getCaseDetails());
         ccdCallbackDto.setEnableEmailNotification(bundle.getValue().getEnableEmailNotificationAsBoolean());
         Set<ConstraintViolation<CcdBundleDTO>> violations = validator.validate(bundle.getValue());
@@ -88,14 +54,7 @@ public class CcdBundleStitchingService implements CcdCaseUpdater {
             return bundle;
         } catch (InterruptedException e) {
             Thread.currentThread().interrupt();
-            logger.error(String.format("Stitching Failed for caseId : %s with issue : %s ",
-                    StringUtilities.convertValidLog(cdamDto.getCaseId()),
-                    StringUtilities.convertValidLog(e.getMessage())));
             throw new StitchingServiceException(e.getMessage(), e);
         }
-    }
-
-    private CcdValue<CcdBundleDTO> bundleJsonToBundleValue(JsonNode jsonNode) throws IOException {
-        return objectMapper.readValue(objectMapper.treeAsTokens(jsonNode), type);
     }
 }
